backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pylint==3.0.3
pymongo==4.6.1
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
certifi==2024.8.30
